backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Computer Vision and Image Processing
opencv-python==4.8.1.78
numpy>=1.21.0,<2.0.0
Pillow>=10.0.0

# HTTP requests
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Machine Learning and Training
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0

# Development and testing (optional)
pytest==7.4.0
pytest-flask==1.2.0